Move curatingInstitutions from ElevatedResponse to Response
